Split EmptyWalletWindow into BsqEmptyWalletWindow and BtcEmptyWalletWindow
